+++ wire-ios-share-engine/Sources/SharingSession.swift
@@ -555,76 +555,3 @@ struct NullInitiateResetMLSConversationUseCase: WireRequestStrategy.InitiateRese
         // do nothing
     }
 }
-
-// TODO: [WPB-17731] remove this duplication
-private extension BackendEnvironment2 {
-
-    init(_ legacyEnvironment: WireTransport.BackendEnvironment) {
-        let environmentType: EnvironmentType = switch legacyEnvironment.environmentType.value {
-        case .default:
-            .default
-        case .staging:
-            .staging
-        case .anta:
-            .anta
-        case .bella:
-            .bella
-        case .chala:
-            .chala
-        case .diya:
-            .diya
-        case .elna:
-            .elna
-        case .foma:
-            .foma
-        case let .custom(url):
-            .custom(url: url)
-        }
-
-        let endpoints = Endpoints(
-            restAPIURL: legacyEnvironment.backendURL,
-            websocketURL: legacyEnvironment.backendWSURL,
-            blacklistURL: legacyEnvironment.blackListURL,
-            teamsURL: legacyEnvironment.teamsURL,
-            accountsURL: legacyEnvironment.accountsURL,
-            websiteURL: legacyEnvironment.websiteURL,
-            countlyURL: legacyEnvironment.countlyURL
-        )
-
-        let pinnedKeys: [PinnedKey] = legacyEnvironment.trustData.map {
-            PinnedKey(
-                key: $0.certificateKey,
-                rawKey: $0.rawCertificateKey,
-                hosts: $0.hosts.map { host in
-                    switch host.rule {
-                    case .endsWith:
-                        .endsWith(host.value)
-                    case .equals:
-                        .equals(host.value)
-                    }
-                }
-            )
-        }
-
-        let proxyConfig = legacyEnvironment.proxy.map {
-            ProxyConfig(
-                host: $0.host,
-                port: $0.port,
-                needsAuthentication: $0.needsAuthentication
-            )
-        }
-
-        let config = Config(
-            endpoints: endpoints,
-            pinnedKeys: pinnedKeys,
-            proxyConfig: proxyConfig
-        )
-
-        self.init(
-            title: legacyEnvironment.title,
-            environmentType: environmentType,
-            config: config
-        )
-    }
-
-}
