+++ wire-ios-sync-engine/Source/SessionManager/UserSessionLoader.swift
@@ -505,224 +505,3 @@ final class UserSessionLoader {
     }
 
 }
-
-private extension BackendEnvironment2 {
-
-    init(_ legacyEnvironment: WireTransport.BackendEnvironment) {
-        let environmentType: EnvironmentType = switch legacyEnvironment.environmentType.value {
-        case .default:
-            .default
-        case .staging:
-            .staging
-        case .anta:
-            .anta
-        case .bella:
-            .bella
-        case .chala:
-            .chala
-        case .diya:
-            .diya
-        case .elna:
-            .elna
-        case .foma:
-            .foma
-        case let .custom(url):
-            .custom(url: url)
-        }
-
-        let endpoints = Endpoints(
-            restAPIURL: legacyEnvironment.backendURL,
-            websocketURL: legacyEnvironment.backendWSURL,
-            blacklistURL: legacyEnvironment.blackListURL,
-            teamsURL: legacyEnvironment.teamsURL,
-            accountsURL: legacyEnvironment.accountsURL,
-            websiteURL: legacyEnvironment.websiteURL,
-            countlyURL: legacyEnvironment.countlyURL
-        )
-
-        let pinnedKeys: [PinnedKey] = legacyEnvironment.trustData.map {
-            PinnedKey(
-                key: $0.certificateKey,
-                rawKey: $0.rawCertificateKey,
-                hosts: $0.hosts.map { host in
-                    switch host.rule {
-                    case .endsWith:
-                        .endsWith(host.value)
-                    case .equals:
-                        .equals(host.value)
-                    }
-                }
-            )
-        }
-
-        let proxyConfig = legacyEnvironment.proxy.map {
-            ProxyConfig(
-                host: $0.host,
-                port: $0.port,
-                needsAuthentication: $0.needsAuthentication
-            )
-        }
-
-        let config = Config(
-            endpoints: endpoints,
-            pinnedKeys: pinnedKeys,
-            proxyConfig: proxyConfig
-        )
-
-        self.init(
-            title: legacyEnvironment.title,
-            environmentType: environmentType,
-            config: config
-        )
-    }
-
-}
-
-private struct ProxyCredentialStore {
-
-    let keychain = Keychain()
-
-    func fetchCredentials(
-        host: String,
-        port: Int
-    ) async throws -> (username: String, password: String)? {
-        let usernameData: Data? = try await keychain.fetchItem(query: [
-            .itemClass(.genericPassword),
-            .account("proxy-\(host):\(port)-username"),
-            .returningData(true)
-        ])
-
-        let passwordData: Data? = try await keychain.fetchItem(query: [
-            .itemClass(.genericPassword),
-            .account("proxy-\(host):\(port)-password"),
-            .returningData(true)
-        ])
-
-        guard
-            let usernameData,
-            let passwordData
-        else {
-            return nil
-        }
-
-        return (
-            username: String(decoding: usernameData, as: UTF8.self),
-            password: String(decoding: passwordData, as: UTF8.self)
-        )
-    }
-
-}
-
-private extension WireNetwork.APIVersion {
-
-    init(_ legacyVersion: WireTransport.APIVersion) {
-        switch legacyVersion {
-        case .v0:
-            self = .v0
-        case .v1:
-            self = .v1
-        case .v2:
-            self = .v2
-        case .v3:
-            self = .v3
-        case .v4:
-            self = .v4
-        case .v5:
-            self = .v5
-        case .v6:
-            self = .v6
-        case .v7:
-            self = .v7
-        case .v8:
-            self = .v8
-        case .v9:
-            self = .v9
-        case .v10:
-            self = .v10
-        case .v11:
-            self = .v11
-        }
-    }
-
-}
-
-extension WireTransport.BackendEnvironment {
-
-    convenience init(_ backendEnvironment: BackendEnvironment2) {
-        let trustData: [TrustData] = backendEnvironment.config.pinnedKeys.map { pinnedKey in
-            TrustData(
-                certificateKey: pinnedKey.key,
-                rawCertificateKey: pinnedKey.rawKey,
-                hosts: pinnedKey.hosts.map { host in
-                    switch host {
-                    case let .endsWith(value):
-                        TrustData.Host(
-                            rule: .endsWith,
-                            value: value
-                        )
-                    case let .equals(value):
-                        TrustData.Host(
-                            rule: .equals,
-                            value: value
-                        )
-                    }
-                }
-            )
-        }
-
-        let environmentType: EnvironmentType = switch backendEnvironment.environmentType {
-        case .default:
-            .default
-        case .staging:
-            .staging
-        case .anta:
-            .anta
-        case .bella:
-            .bella
-        case .chala:
-            .chala
-        case .diya:
-            .diya
-        case .elna:
-            .elna
-        case .foma:
-            .foma
-        case let .custom(url):
-            .custom(url: url)
-        }
-
-        let endpoints = BackendEndpoints(
-            backendURL: backendEnvironment.config.endpoints.restAPIURL,
-            backendWSURL: backendEnvironment.config.endpoints.websocketURL,
-            blackListURL: backendEnvironment.config.endpoints.blacklistURL,
-            teamsURL: backendEnvironment.config.endpoints.teamsURL,
-            accountsURL: backendEnvironment.config.endpoints.accountsURL,
-            websiteURL: backendEnvironment.config.endpoints.websiteURL,
-            countlyURL: backendEnvironment.config.endpoints.countlyURL
-        )
-
-        var proxySettings: WireTransport.ProxySettings?
-        if let proxyConfig = backendEnvironment.config.proxyConfig {
-            proxySettings = WireTransport.ProxySettings(
-                host: proxyConfig.host,
-                port: proxyConfig.port,
-                needsAuthentication: proxyConfig.needsAuthentication
-            )
-        }
-
-        let certificateTrust = ServerCertificateTrust(
-            trustData: trustData,
-            currentDateProvider: .system
-        )
-
-        self.init(
-            title: backendEnvironment.title,
-            trustData: trustData,
-            environmentType: environmentType,
-            endpoints: endpoints,
-            proxySettings: proxySettings,
-            certificateTrust: certificateTrust
-        )
-    }
-
-}
